final_project/requirements.txt
python-dateutil==2.8.0
python-dotenv==0.10.3
requests==2.22.0
requests-oauthlib==1.2.0
tweepy==3.8.0
urllib3==1.25.3
